# backend/docs/swagger.yaml
definitions:
  domain.CategoryParams:
    properties:
      category:
        type: integer
      pattern:
        type: string
    type: object
...
  domain.GetUserInfoStruct:
    properties:
      access:
        type: integer
      ban_reason:
        type: string
      ban_time:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  domain.GetUserStatsRes:
    properties:
      user_stats:
        $ref: '#/definitions/domain.PlayerStats'
    type: object
  domain.GetUsersTop:
    properties:
      users:
        items:
          $ref: '#/definitions/domain.PlayerStats'
        type: array
    type: object
...
  domain.LevelComplaint:
    properties:
      author_id:
        type: integer
      level_id:
        type: integer
      message:
        type: string
      reason:
        type: string
    required:
    - author_id
    - level_id
    - message
    - reason
    type: object
...
  domain.PlayerStats:
    properties:
      average_accuracy_classic:
        type: number
      average_accuracy_relax:
        type: number
      average_delay:
        type: number
      num_chars_classic:
        type: integer
      num_chars_relax:
        type: integer
      num_classes_classic:
        items:
          type: integer
        type: array
      num_games_mult:
        type: integer
      num_level_classic:
        type: integer
      num_level_relax:
        type: integer
      num_press_err_by_char_by_lang:
        additionalProperties:
          additionalProperties:
            items:
              type: integer
            type: array
          type: object
        description: NumPressErrByCharByLang map[string]interface{} `json:"num_press_err_by_char_by_lang"
          db:"num_press_err_by_char_by_lang" binding:"required"`
        type: object
      sum_points:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
      win_percentage:
        type: number
    required:
    - num_press_err_by_char_by_lang
    type: object
...
  domain.StatSortFilterParams:
    properties:
      category_params:
        $ref: '#/definitions/domain.CategoryParams'
      page_info:
        $ref: '#/definitions/github_com_Gadzet005_GoType_backend_internal_domain_Statistics.PageInfo'
      points:
        type: string
    required:
    - page_info
    type: object
  domain.User:
    properties:
      name:
        type: string
      password:
        description: hash
        type: string
    required:
    - name
    - password
    type: object
...
  domain.UserComplaint:
    properties:
      author_id:
        type: integer
      message:
        type: string
      reason:
        type: string
      user_id:
        type: integer
    required:
    - author_id
    - message
    - reason
    - user_id
    type: object
...
  domain.UserID:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  domain.UserSearchParams:
    properties:
      is_banned:
        type: boolean
      name:
        type: string
      offset:
        type: integer
      page_size:
        type: integer
    required:
    - is_banned
    - name
    - offset
    - page_size
    type: object
...
  domain.Users:
    properties:
      users:
        items:
          $ref: '#/definitions/domain.User'
        type: array
    type: object
...
  github_com_Gadzet005_GoType_backend_internal_domain_Statistics.PageInfo:
    properties:
      offset:
        type: integer
      page_size:
        type: integer
    required:
    - offset
    - page_size
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for GoType game and website
  title: GoType App API
  version: 0.0.1
paths:
...
  /admin/get-users:
    get:
      consumes:
      - application/json
      description: Get list of users with given params. Available only for moderators
        and admins
      operationId: get-users
      parameters:
      - description: Search params - name, isBanned, pages info
        in: body
        name: user_search_params
        required: true
        schema:
          $ref: '#/definitions/domain.UserSearchParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Users'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Users with given params
      tags:
      - admin
...
  /stats/get-user-stats:
    get:
      consumes:
      - application/json
      description: Get User Statistics by id
      operationId: get-user-stats
      parameters:
      - description: Id of user
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UserID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetUserStatsRes'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_NO_SUCH_USER - There is no user with
            such id'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get User Statistics
      tags:
      - stats
  /stats/get-users-top:
    get:
      consumes:
      - application/json
      description: Get Users Top with given params
      operationId: get-users-top
      parameters:
      - description: Search params
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.StatSortFilterParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetUsersTop'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_NO_SUCH_USER - There is no user with
            such id'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Users Top
      tags:
      - stats
  /user-actions/change-avatar:
    post:
      consumes:
      - multipart/form-data
      description: Change avatar. In case of empty request sets default avatar.
      operationId: change-avatar
      parameters:
      - description: File with new avatar image
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input files (e.g. wrong name, not "avatar");'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Change Avatar
      tags:
      - user-actions
  /user-actions/get-user-info:
    get:
      consumes:
      - application/json
      description: Get username by id
      operationId: get-user-info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetUserInfoStruct'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - There is no id
            in token payload/Wrong structure of Access Token/No Access Token; ERR_NO_SUCH_USER
            - User with such id not found;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get User Info
      tags:
      - user-actions
  /user-actions/logout:
    post:
      consumes:
      - application/json
      description: Expire refreshToken manually
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - There is no id
            in token payload/Wrong structure of Access Token/No Access Token; ERR_NO_SUCH_USER
            - User with such id not found;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Logout
      tags:
      - user-actions
  /user-actions/write-level-complaint:
    post:
      consumes:
      - application/json
      description: 'Send level complaint to server. Possible Reason values: Offencive
        name, Offencive video, Offencive audio, Offencive text'
      operationId: write-level-complaint
      parameters:
      - description: new complaint info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.LevelComplaint'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input json/Invalid Reason;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Write Level Complaint
      tags:
      - user-actions
  /user-actions/write-user-complaint:
    post:
      consumes:
      - application/json
      description: 'Send user complaint to server. Possible Reason values: Cheating,
        Offencive nickname, Unsportsmanlike conduct'
      operationId: write-user-complaint
      parameters:
      - description: new complaint info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UserComplaint'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input json/Invalid Reason;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Write User Complaint
      tags:
      - user-actions
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"